+++ source/isaaclab/isaaclab/sim/simulation_context.py
@@ -32,11 +32,14 @@ from isaacsim.core.utils.viewports import set_camera_view
 from isaacsim.core.version import get_version
 from pxr import Gf, PhysxSchema, Sdf, Usd, UsdPhysics
 
-from isaaclab.sim.utils import stage as stage_utils
+import isaaclab.sim.utils.stage as stage_utils
 
 from .simulation_cfg import SimulationCfg
 from .spawners import DomeLightCfg, GroundPlaneCfg
 from .utils import ColoredFormatter, RateLimitFilter, bind_physics_material
+
+# import logger
+logger = logging.getLogger(__name__)
 
 
 class SimulationContext(_SimulationContext):
@@ -246,7 +249,7 @@ class SimulationContext(_SimulationContext):
             )
         else:
             self._app_control_on_stop_handle = None
-        self._disable_app_control_on_stop_handle = False
+        self._disable_app_control_on_stop_handle = True
 
         # flatten out the simulation dictionary
         sim_params = self.cfg.to_dict()
@@ -539,7 +542,7 @@ class SimulationContext(_SimulationContext):
         if self._anim_recording_enabled:
             is_anim_recording_finished = self._update_anim_recording()
             if is_anim_recording_finished:
-                carb.log_warn("[INFO][SimulationContext]: Animation recording finished. Closing app.")
+                logger.warning("[INFO][SimulationContext]: Animation recording finished. Closing app.")
                 self._app.shutdown()
 
         # check if the simulation timeline is paused. in that case keep stepping until it is playing
@@ -927,7 +930,7 @@ class SimulationContext(_SimulationContext):
     def _finish_anim_recording(self):
         """Finishes the animation recording and outputs the baked animation recording."""
 
-        carb.log_warn(
+        logger.warning(
             "[INFO][SimulationContext]: Finishing animation recording. Stage must be saved. Might take a few minutes."
         )
 
@@ -1124,6 +1127,7 @@ def build_simulation_context(
         sim.logger.error(traceback.format_exc())
         raise
     finally:
+        sim._disable_app_control_on_stop_handle = True
         if not sim.has_gui():
             # Stop simulation only if we aren't rendering otherwise the app will hang indefinitely
             sim.stop()