+++ source/isaaclab/isaaclab/sim/spawners/wrappers/wrappers_cfg.py
@@ -34,11 +34,20 @@ class MultiAssetSpawnerCfg(RigidObjectSpawnerCfg, DeformableObjectSpawnerCfg):
     assets_cfg: list[SpawnerCfg] = MISSING
     """List of asset configurations to spawn."""
 
-    random_choice: bool = True
-    """Whether to randomly select an asset configuration. Default is True.
+    enable_clone: bool = False
+    """This is enables the cloning of spawned assets from first instance to all other instances that has the
+    same prefix path pattern, for example, given /World/env_.*/asset_.*, multi-asset spawner will spawn asset_0,
+    asset_1, asset_n under /World/env_0. Then if enable_clone is set to True, env_0 is cloned to env_1, env_2, etc.
+    If False, .* is not allowed in the prefix path of the prim_path, an safty check error will be raised."""
 
-    If False, the asset configurations are spawned in the order they are provided in the list.
-    If True, a random asset configuration is selected for each spawn.
+    random_choice: bool = True
+    """ This parameter is ignored.
+    See :attr:`isaaclab.scene.interactive_scene_cfg.InteractiveSceneCfg.random_heterogeneous_cloning` for details.
+
+    .. warning::
+
+        This attribute is deprecated. Use
+        :attr:`~isaaclab.scene.interactive_scene_cfg.InteractiveSceneCfg.random_heterogeneous_cloning` instead.
     """
 
 
@@ -64,4 +73,9 @@ class MultiUsdFileCfg(UsdFileCfg):
 
     If False, the asset configurations are spawned in the order they are provided in the list.
     If True, a random asset configuration is selected for each spawn.
+
+    .. warning::
+
+        This attribute is deprecated. Use
+        :attr:`~isaaclab.scene.interactive_scene_cfg.InteractiveSceneCfg.random_heterogeneous_cloning` instead.
     """